ci: ensure global tests cleanup
