+++ test/integration/promise-wrappers/test-promise-wrappers.test.mts
@@ -10,7 +10,6 @@ import { config } from '../../common.test.mjs';
 type TttRow = RowDataPacket & { ttt: number };
 type QqqRow = RowDataPacket & { qqq: number };
 type TttUuuRow = RowDataPacket & { ttt: number; uuu: string };
-type CurrentUserRow = RowDataPacket & { 'current_user()': string };
 
 if (`${process.env.MYSQL_CONNECTION_URL}`.includes('pscale_pw_')) {
   skip('Skipping test for PlanetScale');
@@ -277,57 +276,6 @@ await describe('Promise Wrappers', async () => {
         eventName
       );
     }
-  });
-
-  await it('testChangeUser', async () => {
-    const onlyUsername = function (name: string) {
-      return name.substring(0, name.indexOf('@'));
-    };
-
-    const conn = await createConnection(config);
-    await conn.query(
-      "CREATE USER IF NOT EXISTS 'changeuser1'@'%' IDENTIFIED BY 'changeuser1pass'"
-    );
-    await conn.query(
-      "CREATE USER IF NOT EXISTS 'changeuser2'@'%' IDENTIFIED BY 'changeuser2pass'"
-    );
-    await conn.query("GRANT ALL ON *.* TO 'changeuser1'@'%'");
-    await conn.query("GRANT ALL ON *.* TO 'changeuser2'@'%'");
-    await conn.query('FLUSH PRIVILEGES');
-
-    await conn.changeUser({
-      user: 'changeuser1',
-      password: 'changeuser1pass',
-    });
-    const result1 = await conn.query<CurrentUserRow[]>('select current_user()');
-    strict.deepEqual(
-      onlyUsername(result1[0][0]['current_user()']),
-      'changeuser1'
-    );
-
-    await conn.changeUser({
-      user: 'changeuser2',
-      password: 'changeuser2pass',
-    });
-    const result2 = await conn.query<CurrentUserRow[]>('select current_user()');
-    strict.deepEqual(
-      onlyUsername(result2[0][0]['current_user()']),
-      'changeuser2'
-    );
-
-    await conn.changeUser({
-      user: 'changeuser1',
-      // TODO: re-enable testing passwordSha1 auth. Only works for mysql_native_password, so need to change test to create user with this auth method
-      password: 'changeuser1pass',
-      //passwordSha1: Buffer.from('f961d39c82138dcec42b8d0dcb3e40a14fb7e8cd', 'hex') // sha1(changeuser1pass)
-    });
-    const result3 = await conn.query<CurrentUserRow[]>('select current_user()');
-    strict.deepEqual(
-      onlyUsername(result3[0][0]['current_user()']),
-      'changeuser1'
-    );
-
-    await conn.end();
   });
 
   await it('testConnectionProperties', async () => {
